chore: row, select, textarea, point, radio, and number fields
