+++ packages/ui/src/forms/field-types/Number/index.tsx
@@ -1,62 +1,128 @@
-import React from 'react'
+'use client'
+import React, { useCallback, useEffect, useState } from 'react'
 
 import type { Props } from './types'
 import { isNumber } from 'payload/utilities'
 import ReactSelect from '../../../elements/ReactSelect'
-import DefaultError from '../../Error'
-import FieldDescription from '../../FieldDescription'
-import DefaultLabel from '../../Label'
-import { NumberInput } from './Input'
-import { NumberInputWrapper } from './Wrapper'
 import { withCondition } from '../../withCondition'
+import { fieldBaseClass } from '../shared'
+import { getTranslation } from '@payloadcms/translations'
+import { useTranslation } from '../../../providers/Translation'
+import useField from '../../useField'
+import { Option } from '../../../elements/ReactSelect/types'
 
 import './index.scss'
 
 const NumberField: React.FC<Props> = (props) => {
   const {
     name,
-    admin: {
-      className,
-      components: { Error, Label, afterInput, beforeInput } = {},
-      description,
-      placeholder,
-      readOnly,
-      step,
-      style,
-      width,
-    } = {},
-    hasMany,
-    label,
-    maxRows,
-    min,
+    className,
+    placeholder,
+    readOnly,
+    style,
+    width,
     path: pathFromProps,
     required,
-    valid = true,
-    i18n,
-    value,
+    Error,
+    Label,
+    Description,
+    BeforeInput,
+    AfterInput,
+    validate,
   } = props
 
-  const ErrorComp = Error || DefaultError
-  const LabelComp = Label || DefaultLabel
+  const max = 'max' in props ? props.max : Infinity
+  const min = 'min' in props ? props.min : -Infinity
+  const step = 'step' in props ? props.step : 1
+  const hasMany = 'hasMany' in props ? props.hasMany : false
+  const maxRows = 'maxRows' in props ? props.maxRows : Infinity
 
-  const path = pathFromProps || name
+  const { i18n } = useTranslation()
+
+  const memoizedValidate = useCallback(
+    (value, options) => {
+      if (typeof validate === 'function') return validate(value, { ...options, max, min, required })
+    },
+    [validate, min, max, required],
+  )
+
+  const { setValue, showError, value, path } = useField<number | number[]>({
+    path: pathFromProps || name,
+    validate: memoizedValidate,
+  })
+
+  const handleChange = useCallback(
+    (e) => {
+      const val = parseFloat(e.target.value)
+
+      if (Number.isNaN(val)) {
+        setValue('')
+      } else {
+        setValue(val)
+      }
+    },
+    [setValue],
+  )
+
+  const [valueToRender, setValueToRender] = useState<
+    { id: string; label: string; value: { value: number } }[]
+  >([]) // Only for hasMany
+
+  const handleHasManyChange = useCallback(
+    (selectedOption) => {
+      if (!readOnly) {
+        let newValue
+        if (!selectedOption) {
+          newValue = []
+        } else if (Array.isArray(selectedOption)) {
+          newValue = selectedOption.map((option) => Number(option.value?.value || option.value))
+        } else {
+          newValue = [Number(selectedOption.value?.value || selectedOption.value)]
+        }
+
+        setValue(newValue)
+      }
+    },
+    [readOnly, setValue],
+  )
+
+  // useEffect update valueToRender:
+  useEffect(() => {
+    if (hasMany && Array.isArray(value)) {
+      setValueToRender(
+        value.map((val, index) => {
+          return {
+            id: `${val}${index}`, // append index to avoid duplicate keys but allow duplicate numbers
+            label: `${val}`,
+            value: {
+              toString: () => `${val}${index}`,
+              value: (val as any)?.value || val,
+            }, // You're probably wondering, why the hell is this done that way? Well, React-select automatically uses "label-value" as a key, so we will get that react duplicate key warning if we just pass in the value as multiple values can be the same. So we need to append the index to the toString() of the value to avoid that warning, as it uses that as the key.
+          }
+        }),
+      )
+    }
+  }, [value, hasMany])
 
   return (
-    <NumberInputWrapper
-      className={className}
-      readOnly={readOnly}
-      hasMany={hasMany}
-      style={style}
-      width={width}
-      path={path}
+    <div
+      className={[
+        fieldBaseClass,
+        'number',
+        className,
+        showError && 'error',
+        readOnly && 'read-only',
+        hasMany && 'has-many',
+      ]
+        .filter(Boolean)
+        .join(' ')}
+      style={{
+        ...style,
+        width,
+      }}
     >
-      <ErrorComp path={path} />
-      <LabelComp
-        htmlFor={`field-${path.replace(/\./g, '__')}`}
-        label={label}
-        required={required}
-        i18n={i18n}
-      />
+      {Error}
+      {Label}
       {hasMany ? (
         <ReactSelect
           className={`field-${path.replace(/\./g, '__')}`}
@@ -81,27 +147,32 @@ const NumberField: React.FC<Props> = (props) => {
           // onChange={handleHasManyChange}
           options={[]}
           // placeholder={t('general:enterAValue')}
-          showError={!valid}
-          // value={valueToRender as Option[]}
+          showError={showError}
+          value={valueToRender as Option[]}
         />
       ) : (
         <div>
-          {Array.isArray(beforeInput) && beforeInput.map((Component, i) => <Component key={i} />)}
-          <NumberInput
-            path={path}
-            required={required}
-            min={min}
-            placeholder={placeholder}
-            readOnly={readOnly}
+          {BeforeInput}
+          <input
+            disabled={readOnly}
+            id={`field-${path.replace(/\./g, '__')}`}
+            name={path}
+            onChange={handleChange}
+            onWheel={(e) => {
+              // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+              // @ts-ignore
+              e.target.blur()
+            }}
+            placeholder={getTranslation(placeholder, i18n)}
             step={step}
-            hasMany={hasMany}
-            name={name}
+            type="number"
+            value={typeof value === 'number' ? value : ''}
           />
-          {Array.isArray(afterInput) && afterInput.map((Component, i) => <Component key={i} />)}
+          {AfterInput}
         </div>
       )}
-      <FieldDescription description={description} value={value} i18n={i18n} />
-    </NumberInputWrapper>
+      {Description}
+    </div>
   )
 }
