+++ packages/ui/src/forms/field-types/Number/types.ts
@@ -1,7 +1,7 @@
 import type { NumberField } from 'payload/types'
 import type { FormFieldBase } from '../shared'
 
-export type Props = FormFieldBase &
-  Omit<NumberField, 'type'> & {
-    value?: number
-  }
+export type Props = FormFieldBase & {
+  path?: string
+  name?: string
+}
