+++ packages/ui/src/forms/field-types/Point/index.tsx
@@ -1,13 +1,12 @@
-import React from 'react'
+import React, { useCallback } from 'react'
 
 import type { Props } from './types'
 
 import { getTranslation } from '@payloadcms/translations'
-import DefaultError from '../../Error'
-import FieldDescription from '../../FieldDescription'
-import DefaultLabel from '../../Label'
-import { NumberInputWrapper } from '../Number/Wrapper'
-import { PointInput } from './Input'
+import useField from '../../useField'
+import { fieldBaseClass } from '../shared'
+import { Validate } from 'payload/types'
+import { useTranslation } from '../../../providers/Translation'
 
 import './index.scss'
 
@@ -16,81 +15,104 @@ const baseClass = 'point'
 const PointField: React.FC<Props> = (props) => {
   const {
     name,
-    admin: {
-      className,
-      components: { Error, Label, afterInput, beforeInput } = {},
-      description,
-      placeholder,
-      readOnly,
-      step,
-      style,
-      width,
-    } = {},
-    label,
-    path: pathFromProps,
+    className,
+    placeholder,
+    readOnly,
+    style,
+    width,
     required,
-    i18n,
-    i18n: { t },
-    value,
+    validate,
+    path: pathFromProps,
+    Error,
+    BeforeInput,
+    AfterInput,
+    Label,
+    Description,
   } = props
 
-  const ErrorComp = Error || DefaultError
-  const LabelComp = Label || DefaultLabel
+  const { i18n } = useTranslation()
 
-  const path = pathFromProps || name
+  const step = 'step' in props ? props.step : 1
+
+  const memoizedValidate: Validate = useCallback(
+    (value, options) => {
+      if (typeof validate === 'function') return validate(value, { ...options, required })
+    },
+    [validate, required],
+  )
+
+  const {
+    setValue,
+    value = [null, null],
+    showError,
+    path,
+  } = useField<[number, number]>({
+    validate: memoizedValidate,
+    path: pathFromProps || name,
+  })
+
+  const handleChange = useCallback(
+    (e, index: 0 | 1) => {
+      let val = parseFloat(e.target.value)
+      if (Number.isNaN(val)) {
+        val = e.target.value
+      }
+      const coordinates = [...value]
+      coordinates[index] = val
+      setValue(coordinates)
+    },
+    [setValue, value],
+  )
 
   return (
-    <NumberInputWrapper
-      className={[className, baseClass].filter(Boolean).join(' ')}
-      readOnly={readOnly}
-      style={style}
-      width={width}
-      path={path}
+    <div
+      className={[fieldBaseClass, className, showError && 'error', readOnly && 'read-only']
+        .filter(Boolean)
+        .join(' ')}
+      style={{
+        ...style,
+        width,
+      }}
     >
-      <ErrorComp path={path} />
+      {Error}
       <ul className={`${baseClass}__wrap`}>
         <li>
-          <LabelComp
-            htmlFor={`field-longitude-${path.replace(/\./g, '__')}`}
-            label={`${getTranslation(label || name, i18n)} - ${t('fields:longitude')}`}
-            required={required}
-            i18n={i18n}
-          />
+          {Label}
           <div className="input-wrapper">
-            {Array.isArray(beforeInput) && beforeInput.map((Component, i) => <Component key={i} />)}
-            <PointInput
-              path={path}
-              placeholder={placeholder}
-              readOnly={readOnly}
+            {BeforeInput}
+            <input
+              disabled={readOnly}
+              id={`field-longitude-${path.replace(/\./g, '__')}`}
+              name={`${path}.longitude`}
+              onChange={(e) => handleChange(e, 0)}
+              placeholder={getTranslation(placeholder, i18n)}
               step={step}
-              required={required}
-              isLatitude={false}
+              type="number"
+              value={value && typeof value[0] === 'number' ? value[0] : ''}
             />
-            {Array.isArray(afterInput) && afterInput.map((Component, i) => <Component key={i} />)}
+            {AfterInput}
           </div>
         </li>
         <li>
-          <LabelComp
-            htmlFor={`field-latitude-${path.replace(/\./g, '__')}`}
-            label={`${getTranslation(label || name, i18n)} - ${t('fields:latitude')}`}
-            required={required}
-            i18n={i18n}
-          />
+          {Label}
           <div className="input-wrapper">
-            {Array.isArray(beforeInput) && beforeInput.map((Component, i) => <Component key={i} />)}
-            <PointInput
-              path={path}
-              placeholder={placeholder}
-              readOnly={readOnly}
+            {BeforeInput}
+            <input
+              disabled={readOnly}
+              id={`field-latitude-${path.replace(/\./g, '__')}`}
+              name={`${path}.latitude`}
+              onChange={(e) => handleChange(e, 1)}
+              placeholder={getTranslation(placeholder, i18n)}
               step={step}
-              required={required}
+              type="number"
+              value={value && typeof value[1] === 'number' ? value[1] : ''}
             />
-            {Array.isArray(afterInput) && afterInput.map((Component, i) => <Component key={i} />)}
+            {AfterInput}
           </div>
         </li>
       </ul>
-      <FieldDescription description={description} path={path} value={value} i18n={i18n} />
-    </NumberInputWrapper>
+      {Description}
+    </div>
   )
 }
