--- packages/ui/src/forms/field-types/RadioGroup/Wrapper/index.scss
@@ -1,46 +1,0 @@
-@import '../../../../scss/styles.scss';
-
-.radio-group {
-  ul {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-  }
-
-  &--layout-horizontal {
-    ul {
-      display: flex;
-      flex-wrap: wrap;
-    }
-
-    li {
-      flex-shrink: 0;
-      [dir='ltr'] & {
-        padding-right: $baseline;
-      }
-      [dir='rtl'] & {
-        padding-left: $baseline;
-      }
-    }
-  }
-}
-
-html[data-theme='light'] {
-  .radio-group {
-    &.error {
-      .radio-input__styled-radio {
-        @include lightInputError;
-      }
-    }
-  }
-}
-
-html[data-theme='dark'] {
-  .radio-group {
-    &.error {
-      .radio-input__styled-radio {
-        @include darkInputError;
-      }
-    }
-  }
-}
