+++ packages/ui/src/forms/field-types/RadioGroup/index.scss
@@ -4,4 +4,47 @@
   &__error-wrap {
     position: relative;
   }
+
+  &--layout-horizontal {
+    ul {
+      display: flex;
+      flex-wrap: wrap;
+    }
+
+    li {
+      flex-shrink: 0;
+      [dir='ltr'] & {
+        padding-right: $baseline;
+      }
+      [dir='rtl'] & {
+        padding-left: $baseline;
+      }
+    }
+  }
+
+  ul {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+  }
+}
+
+html[data-theme='light'] {
+  .radio-group {
+    &.error {
+      .radio-input__styled-radio {
+        @include lightInputError;
+      }
+    }
+  }
+}
+
+html[data-theme='dark'] {
+  .radio-group {
+    &.error {
+      .radio-input__styled-radio {
+        @include darkInputError;
+      }
+    }
+  }
 }
