+++ packages/ui/src/forms/field-types/RadioGroup/types.ts
@@ -1,10 +1,9 @@
 import type { RadioField } from 'payload/types'
 import { FormFieldBase } from '../shared'
 
-export type Props = FormFieldBase &
-  Omit<RadioField, 'type'> & {
-    path?: string
-    value?: string
-  }
+export type Props = FormFieldBase & {
+  name?: string
+  path?: string
+}
 
 export type OnChange<T = string> = (value: T) => void
