+++ packages/ui/src/forms/field-types/Row/index.tsx
@@ -1,8 +1,8 @@
+'use client'
 import React from 'react'
 
 import type { Props } from './types'
 
-import { createNestedFieldPath } from '../../Form/createNestedFieldPath'
 import RenderFields from '../../RenderFields'
 import { fieldBaseClass } from '../shared'
 import { RowProvider } from './provider'
@@ -15,17 +15,10 @@ const baseClass = 'row'
 const Row: React.FC<Props> = (props) => {
   const {
     admin: { className, readOnly },
-    fieldTypes,
-    fields,
     forceRender = false,
     indexPath,
-    path,
     permissions,
-    formState,
-    user,
-    i18n,
-    payload,
-    config,
+    fieldMap,
   } = props
 
   return (
@@ -33,21 +26,12 @@ const Row: React.FC<Props> = (props) => {
       <div className={[fieldBaseClass, baseClass, className].filter(Boolean).join(' ')}>
         <RenderFields
           className={`${baseClass}__fields`}
-          fieldSchema={fields.map((field) => ({
-            ...field,
-            path: createNestedFieldPath(path, field),
-          }))}
-          fieldTypes={fieldTypes}
+          fieldMap={fieldMap}
           forceRender={forceRender}
           indexPath={indexPath}
           margins={false}
           permissions={permissions}
           readOnly={readOnly}
-          formState={formState}
-          user={user}
-          i18n={i18n}
-          payload={payload}
-          config={config}
         />
       </div>
     </RowProvider>
