+++ packages/ui/src/forms/field-types/Select/types.ts
@@ -1,8 +1,7 @@
-import type { SelectField } from 'payload/types'
 import { FormFieldBase } from '../shared'
 
-export type Props = FormFieldBase &
-  Omit<SelectField, 'type'> & {
-    path?: string
-    value?: string
-  }
+export type Props = FormFieldBase & {
+  path?: string
+  value?: string
+  name?: string
+}
