+++ packages/ui/src/forms/field-types/Textarea/types.ts
@@ -1,8 +1,6 @@
-import type { TextareaField } from 'payload/types'
 import { FormFieldBase } from '../shared'
 
-export type Props = FormFieldBase &
-  Omit<TextareaField, 'type'> & {
-    path?: string
-    value: string
-  }
+export type Props = FormFieldBase & {
+  path?: string
+  name?: string
+}
